fix: update changes to original component
