+++ public/pages/Correlations/containers/CorrelationsContainer.tsx
@@ -39,6 +39,13 @@ import {
   EuiBadge,
   EuiFilterGroup,
   EuiHorizontalRule,
+  EuiButtonGroup,
+  EuiBasicTableColumn,
+  EuiToolTip,
+  EuiInMemoryTable,
+  EuiTextColor,
+  EuiLink,
+  EuiFieldSearch,
 } from '@elastic/eui';
 import { FilterItem, FilterGroup } from '../components/FilterGroup';
 import {
@@ -56,8 +63,31 @@ import { renderToStaticMarkup } from 'react-dom/server';
 import { Network } from 'react-graph-vis';
 import { getLogTypeLabel } from '../../LogTypes/utils/helpers';
 import { NotificationsStart } from 'opensearch-dashboards/public';
-import { errorNotificationToast, setBreadcrumbs } from '../../../utils/helpers';
+import {
+  capitalizeFirstLetter,
+  errorNotificationToast,
+  renderVisualization,
+  setBreadcrumbs,
+} from '../../../utils/helpers';
 import { PageHeader } from '../../../components/PageHeader/PageHeader';
+import moment from 'moment';
+import { ChartContainer } from '../../../components/Charts/ChartContainer';
+import {
+  addInteractiveLegends,
+  DateOpts,
+  defaultDateFormat,
+  defaultScaleDomain,
+  defaultTimeUnit,
+  getChartTimeUnit,
+  getDomainRange,
+  getTimeTooltip,
+  getVisualizationSpec,
+  getXAxis,
+  getYAxis,
+} from '../../Overview/utils/helpers';
+import { debounce } from 'lodash';
+
+export const DEFAULT_EMPTY_DATA = '-';
 
 interface CorrelationsProps
   extends RouteComponentProps<
@@ -77,6 +107,23 @@ interface SpecificFindingCorrelations {
   correlatedFindings: CorrelationFinding[];
 }
 
+interface CorrelationsTableData {
+  id: string;
+  startTime: number;
+  correlationRule: string;
+  alertSeverity: string[];
+  logTypes: string[];
+  findingsSeverity: string[];
+  correlatedFindings: CorrelationFinding[];
+}
+
+interface FlyoutTableData {
+  timestamp: string;
+  mitreTactic: string[];
+  detectionRule: string;
+  severity: string;
+}
+
 interface CorrelationsState {
   recentlyUsedRanges: any[];
   graphData: CorrelationGraphData;
@@ -84,6 +131,29 @@ interface CorrelationsState {
   logTypeFilterOptions: FilterItem[];
   severityFilterOptions: FilterItem[];
   loadingGraphData: boolean;
+  isGraphView: Boolean;
+  correlationsTableData: CorrelationsTableData[];
+  connectedFindings: CorrelationFinding[][];
+  isFlyoutOpen: boolean;
+  selectedTableRow: CorrelationsTableData | null;
+  searchTerm: string;
+}
+
+export const renderTime = (time: number | string) => {
+  const momentTime = moment(time);
+  if (time && momentTime.isValid()) return momentTime.format('MM/DD/YY h:mm a');
+  return DEFAULT_EMPTY_DATA;
+};
+
+export interface CorrelationsTableProps {
+  finding: FindingItemType;
+  correlatedFindings: CorrelationFinding[];
+  history: RouteComponentProps['history'];
+  isLoading: boolean;
+  filterOptions: {
+    logTypes: Set<string>;
+    ruleSeverity: Set<string>;
+  };
 }
 
 export class Correlations extends React.Component<CorrelationsProps, CorrelationsState> {
@@ -98,6 +168,12 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
       severityFilterOptions: [...defaultSeverityFilterItemOptions],
       specificFindingInfo: undefined,
       loadingGraphData: false,
+      isGraphView: true,
+      correlationsTableData: [],
+      connectedFindings: [],
+      isFlyoutOpen: false,
+      selectedTableRow: null,
+      searchTerm: '',
     };
   }
 
@@ -123,6 +199,7 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
     setBreadcrumbs([BREADCRUMBS.CORRELATIONS]);
     this.updateState(true /* onMount */);
     this.props.onMount();
+    this.fetchCorrelationsTableData();
   }
 
   componentDidUpdate(
@@ -395,15 +472,74 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
   };
 
   onLogTypeFilterChange = (items: FilterItem[]) => {
-    this.setState({ logTypeFilterOptions: items });
+    this.setState(
+      {
+        logTypeFilterOptions: items,
+      },
+      () => {
+        // If there's specific finding info, update the graph
+        if (this.state.specificFindingInfo) {
+          this.updateGraphDataState(this.state.specificFindingInfo);
+        }
+        // Force update to refresh the table with new filters
+        this.forceUpdate();
+      }
+    );
   };
 
   onSeverityFilterChange = (items: FilterItem[]) => {
-    this.setState({ severityFilterOptions: items });
+    this.setState(
+      {
+        severityFilterOptions: items,
+      },
+      () => {
+        // If there's specific finding info, update the graph
+        if (this.state.specificFindingInfo) {
+          this.updateGraphDataState(this.state.specificFindingInfo);
+        }
+        // Force update to refresh the table with new filters
+        this.forceUpdate();
+      }
+    );
   };
 
   closeFlyout = () => {
     this.setState({ specificFindingInfo: undefined });
+  };
+
+  private openTableFlyout = (correlationTableRow: CorrelationsTableData) => {
+    let newGraphData: CorrelationGraphData = {
+      graph: {
+        nodes: [],
+        edges: [],
+      },
+      events: {
+        click: this.onNodeClick,
+      },
+    };
+
+    if (correlationTableRow.correlatedFindings) {
+      const correlationPairs = this.getCorrelationPairs(correlationTableRow.correlatedFindings);
+      newGraphData = this.prepareGraphData(correlationPairs);
+    }
+
+    // Set all required state at once
+    this.setState({
+      isFlyoutOpen: true,
+      selectedTableRow: correlationTableRow,
+      graphData: newGraphData,
+    });
+  };
+
+  private closeTableFlyout = () => {
+    this.setState({
+      isFlyoutOpen: false,
+      selectedTableRow: null,
+      graphData: {
+        graph: { nodes: [], edges: [] },
+        events: { click: this.onNodeClick },
+      },
+    });
   };
 
   onFindingInspect = async (id: string, logType: string) => {
@@ -440,13 +576,30 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
 
   renderCorrelationsGraph(loadingData: boolean) {
     return this.state.graphData.graph.nodes.length > 0 || loadingData ? (
-      <CorrelationGraph
-        loadingData={loadingData}
-        graph={this.state.graphData.graph}
-        options={{ ...graphRenderOptions }}
-        events={this.state.graphData.events}
-        getNetwork={this.setNetwork}
-      />
+      <>
+        <EuiFlexGroup wrap={true} gutterSize="m" justifyContent="flexStart" alignItems="center">
+          <EuiFlexItem grow={false}>
+            <EuiText size="s">
+              <strong>Severity:</strong>
+            </EuiText>
+          </EuiFlexItem>
+          {ruleSeverity.map((sev, idx) => (
+            <EuiFlexItem grow={false} key={idx}>
+              <EuiText size="s">
+                <EuiIcon type="dot" color={sev.color.background} /> {sev.value}
+              </EuiText>
+            </EuiFlexItem>
+          ))}
+        </EuiFlexGroup>
+        <EuiSpacer />
+        <CorrelationGraph
+          loadingData={loadingData}
+          graph={this.state.graphData.graph}
+          options={{ ...graphRenderOptions }}
+          events={this.state.graphData.events}
+          getNetwork={this.setNetwork}
+        />
+      </>
     ) : (
       <EuiEmptyPrompt
         title={
@@ -467,6 +620,744 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
       />
     );
   }
+
+  private mapConnectedCorrelations(
+    correlations: {
+      finding1: CorrelationFinding;
+      finding2: CorrelationFinding;
+    }[]
+  ) {
+    const connectionsMap = new Map<string, Set<string>>();
+    const findingsMap = new Map<string, CorrelationFinding>();
+
+    correlations.forEach((correlation) => {
+      const { finding1, finding2 } = correlation;
+
+      findingsMap.set(finding1.id, finding1);
+      findingsMap.set(finding2.id, finding2);
+
+      if (!connectionsMap.has(finding1.id)) {
+        connectionsMap.set(finding1.id, new Set<string>());
+      }
+      connectionsMap.get(finding1.id)!.add(finding2.id);
+
+      if (!connectionsMap.has(finding2.id)) {
+        connectionsMap.set(finding2.id, new Set<string>());
+      }
+      connectionsMap.get(finding2.id)!.add(finding1.id);
+    });
+
+    const visited = new Set<string>();
+    const connectedGroups: CorrelationFinding[][] = [];
+
+    function dfs(findingId: string, currentGroup: CorrelationFinding[]) {
+      visited.add(findingId);
+      const finding = findingsMap.get(findingId);
+      if (finding) {
+        currentGroup.push(finding);
+      }
+
+      const connections = connectionsMap.get(findingId) || new Set<string>();
+      connections.forEach((connectedId) => {
+        if (!visited.has(connectedId)) {
+          dfs(connectedId, currentGroup);
+        }
+      });
+    }
+
+    connectionsMap.forEach((_, findingId) => {
+      if (!visited.has(findingId)) {
+        const currentGroup: CorrelationFinding[] = [];
+        dfs(findingId, currentGroup);
+        if (currentGroup.length > 0) {
+          connectedGroups.push(currentGroup);
+        }
+      }
+    });
+
+    return connectedGroups;
+  }
+
+  private fetchCorrelationsTableData = async () => {
+    try {
+      const start = datemath.parse(this.startTime);
+      const end = datemath.parse(this.endTime);
+      const startTime = start?.valueOf() || Date.now();
+      const endTime = end?.valueOf() || Date.now();
+
+      let allCorrelations = await DataStore.correlations.getAllCorrelationsInWindow(
+        startTime.toString(),
+        endTime.toString()
+      );
+
+      const connectedFindings = this.mapConnectedCorrelations(allCorrelations);
+
+      this.setState({ connectedFindings: connectedFindings });
+
+      const tableData: CorrelationsTableData[] = [];
+
+      const allCorrelationRules = await DataStore.correlations.getCorrelationRules();
+      const allCorrelatedAlerts = await DataStore.correlations.getAllCorrelationAlerts();
+
+      const correlationRuleMapsAlerts: { [id: string]: string[] } = {};
+
+      allCorrelationRules.forEach((correlationRule) => {
+        const correlationRuleId = correlationRule.id;
+        correlationRuleMapsAlerts[correlationRuleId] = [];
+
+        allCorrelatedAlerts.correlationAlerts.forEach((correlatedAlert) => {
+          if (correlatedAlert.correlation_rule_id === correlationRuleId) {
+            correlationRuleMapsAlerts[correlationRuleId].push(correlatedAlert.severity);
+          }
+        });
+      });
+
+      for (const findingGroup of connectedFindings) {
+        let correlationRule = '';
+        const logTypes = new Set<string>();
+        const findingsSeverity: string[] = [];
+        let alertsSeverity: string[] = [];
+
+        for (const finding of findingGroup) {
+          findingsSeverity.push(finding.detectionRule.severity);
+          logTypes.add(finding.logType);
+        }
+
+        // Call the APIs only if correlationRule has not been found yet to avoid repeated API calls.
+        if (correlationRule === '') {
+          if (findingGroup[0] && findingGroup[0].detector && findingGroup[0].detector._source) {
+            const correlatedFindingsResponse = await DataStore.correlations.getCorrelatedFindings(
+              findingGroup[0].id,
+              findingGroup[0]?.detector._source?.detector_type
+            );
+            if (
+              correlatedFindingsResponse.correlatedFindings &&
+              correlatedFindingsResponse.correlatedFindings[0] &&
+              correlatedFindingsResponse.correlatedFindings[0].rules
+            ) {
+              const correlationRuleId = correlatedFindingsResponse.correlatedFindings[0].rules[0];
+              const correlationRuleObj =
+                (await DataStore.correlations.getCorrelationRule(correlationRuleId)) || '';
+              alertsSeverity = correlationRuleMapsAlerts[correlationRuleId];
+              if (correlationRuleObj) {
+                correlationRule = correlationRuleObj.name;
+              }
+            }
+          }
+        }
+
+        tableData.push({
+          id: `${startTime}_${findingGroup[0]?.id}`,
+          startTime: startTime,
+          correlationRule: correlationRule,
+          logTypes: Array.from(logTypes),
+          alertSeverity: alertsSeverity,
+          findingsSeverity: findingsSeverity,
+          correlatedFindings: findingGroup,
+        });
+      }
+
+      this.setState({
+        correlationsTableData: tableData,
+      });
+    } catch (error) {
+      console.error('Failed to fetch correlation rules:', error);
+    }
+  };
+
+  private getCorrelatedFindingsVisualizationSpec = (
+    visualizationData: any[],
+    dateOpts: DateOpts = {
+      timeUnit: defaultTimeUnit,
+      dateFormat: defaultDateFormat,
+      domain: defaultScaleDomain,
+    }
+  ) => {
+    return getVisualizationSpec('Correlated Findings data overview', visualizationData, [
+      addInteractiveLegends({
+        mark: {
+          type: 'bar',
+          clip: true,
+        },
+        encoding: {
+          tooltip: [getYAxis('correlatedFinding', 'Correlated Findings'), getTimeTooltip(dateOpts)],
+          x: getXAxis(dateOpts),
+          y: getYAxis('correlatedFinding', 'Count'),
+          color: {
+            field: 'title',
+            legend: {
+              title: 'Legend',
+            },
+          },
+        },
+      }),
+    ]);
+  };
+
+  private generateVisualizationSpec = (connectedFindings: CorrelationFinding[][]) => {
+    const visData = connectedFindings.map((correlatedFindings) => {
+      return {
+        title: 'Correlated Findings',
+        correlatedFinding: correlatedFindings.length,
+        time: correlatedFindings[0].timestamp,
+      };
+    });
+
+    const {
+      dateTimeFilter = {
+        startTime: DEFAULT_DATE_RANGE.start,
+        endTime: DEFAULT_DATE_RANGE.end,
+      },
+    } = this.props;
+
+    const chartTimeUnits = getChartTimeUnit(dateTimeFilter.startTime, dateTimeFilter.endTime);
+
+    return this.getCorrelatedFindingsVisualizationSpec(visData, {
+      timeUnit: chartTimeUnits.timeUnit,
+      dateFormat: chartTimeUnits.dateFormat,
+      domain: getDomainRange(
+        [dateTimeFilter.startTime, dateTimeFilter.endTime],
+        chartTimeUnits.timeUnit.unit
+      ),
+    });
+  };
+
+  private renderCorrelatedFindingsChart = () => {
+    renderVisualization(
+      this.generateVisualizationSpec(this.state.connectedFindings),
+      'correlated-findings-view'
+    );
+
+    return (
+      <>
+        <EuiPanel>
+          <EuiFlexGroup direction="column" gutterSize="m">
+            <EuiFlexItem>
+              <EuiFlexGroup justifyContent="spaceBetween">
+                <EuiFlexItem grow={false}>
+                  <EuiTitle size="s">
+                    <h3>Correlated Findings</h3>
+                  </EuiTitle>
+                </EuiFlexItem>
+              </EuiFlexGroup>
+            </EuiFlexItem>
+            <EuiFlexItem>
+              <ChartContainer chartViewId={'correlated-findings-view'} />
+            </EuiFlexItem>
+          </EuiFlexGroup>
+        </EuiPanel>
+        <EuiSpacer />
+      </>
+    );
+  };
+
+  private getFilteredTableData = (tableData: CorrelationsTableData[]): CorrelationsTableData[] => {
+    const { logTypeFilterOptions, severityFilterOptions } = this.state;
+    const alertSeverityMap: { [key: string]: string } = {
+      '1': 'critical',
+      '2': 'high',
+      '3': 'medium',
+      '4': 'low',
+      '5': 'informational',
+    };
+
+    const selectedLogTypes = logTypeFilterOptions
+      .filter((item) => item.checked === 'on' && item.visible)
+      .map((item) => item.id);
+
+    const selectedSeverities = severityFilterOptions
+      .filter((item) => item.checked === 'on' && item.visible)
+      .map((item) => item.id.toLowerCase());
+
+    return tableData.filter((row) => {
+      const logTypeMatch = row.logTypes.some((logType) => selectedLogTypes.includes(logType));
+
+      const severityMatch = row.alertSeverity.some((severity) =>
+        selectedSeverities.includes(alertSeverityMap[severity])
+      );
+
+      const searchLower = this.state.searchTerm.toLowerCase();
+      const searchMatch =
+        this.state.searchTerm === '' ||
+        row.correlationRule?.toLowerCase().includes(searchLower) ||
+        row.logTypes.some((type) => type.toLowerCase().includes(searchLower)) ||
+        row.alertSeverity.some((severity) =>
+          alertSeverityMap[severity].toLowerCase().includes(searchLower)
+        ) ||
+        row.findingsSeverity.some((severity) => severity.toLowerCase().includes(searchLower));
+      return logTypeMatch && severityMatch && searchMatch;
+    });
+  };
+
+  private debouncedSearch = debounce((searchTerm: string) => {
+    this.setState({ searchTerm }, () => {
+      this.forceUpdate();
+    });
+  }, 300);
+
+  private renderSearchBar = () => {
+    return (
+      <EuiFieldSearch
+        placeholder="Search"
+        value={this.state.searchTerm}
+        onChange={(e) => {
+          e.persist();
+          const searchValue = e.target.value;
+          this.setState({ searchTerm: searchValue });
+          this.debouncedSearch(searchValue);
+        }}
+        fullWidth={true}
+        isClearable={true}
+        compressed={true}
+        aria-label="Search correlations"
+      />
+    );
+  };
+
+  private renderCorrelationsTable = () => {
+    const alertSeverityMap: { [key: string]: string } = {
+      '1': 'critical',
+      '2': 'high',
+      '3': 'medium',
+      '4': 'low',
+      '5': 'informational',
+    };
+
+    const columns: EuiBasicTableColumn<CorrelationsTableData>[] = [
+      {
+        field: 'startTime',
+        name: 'Start time',
+        sortable: true,
+        dataType: 'date',
+        render: (startTime: number) => {
+          return new Date(startTime).toLocaleString();
+        },
+      },
+      {
+        field: 'correlationRule',
+        name: 'Correlation Rule',
+        sortable: true,
+        render: (name: string) => name || 'N/A',
+      },
+      {
+        field: 'logTypes',
+        name: 'Log Types',
+        sortable: true,
+        render: (logTypes: string[]) => {
+          if (!logTypes || logTypes.length === 0) return DEFAULT_EMPTY_DATA;
+          const MAX_DISPLAY = 2;
+          const remainingCount = logTypes.length > MAX_DISPLAY ? logTypes.length - MAX_DISPLAY : 0;
+          const displayedLogTypes = logTypes.slice(0, MAX_DISPLAY).map((logType) => {
+            const label = logType;
+            return <EuiBadge>{label}</EuiBadge>;
+          });
+          const tooltipContent = (
+            <>
+              {logTypes.slice(MAX_DISPLAY).map((logType) => {
+                const label = logType;
+                return (
+                  <div style={{ display: 'flex', flexDirection: 'column', padding: '4px' }}>
+                    <EuiBadge>{label}</EuiBadge>
+                  </div>
+                );
+              })}
+            </>
+          );
+          return (
+            <span
+              style={{
+                display: 'flex',
+                flexDirection: 'row',
+                alignItems: 'center',
+                gap: '4px',
+                whiteSpace: 'normal',
+                flexWrap: 'wrap',
+                width: '100%',
+              }}
+            >
+              {displayedLogTypes}
+              {remainingCount > 0 && (
+                <EuiToolTip content={tooltipContent} position="top">
+                  <EuiBadge>{`+${remainingCount} more`}</EuiBadge>
+                </EuiToolTip>
+              )}
+            </span>
+          );
+        },
+      },
+      {
+        field: 'alertSeverity',
+        name: 'Alert Severity',
+        sortable: true,
+        render: (alertSeverity: string[]) => {
+          if (!alertSeverity || alertSeverity.length === 0) return DEFAULT_EMPTY_DATA;
+          const MAX_DISPLAY = 2;
+          const remainingCount =
+            alertSeverity.length > MAX_DISPLAY ? alertSeverity.length - MAX_DISPLAY : 0;
+          const displayedSeverities = alertSeverity.slice(0, MAX_DISPLAY).map((severity) => {
+            const label = alertSeverityMap[severity];
+            const { background, text } = getSeverityColor(label);
+            return (
+              <EuiBadge key={severity} style={{ backgroundColor: background, color: text }}>
+                {label}
+              </EuiBadge>
+            );
+          });
+
+          const tooltipContent = (
+            <>
+              {alertSeverity.slice(MAX_DISPLAY).map((severity) => {
+                const label = alertSeverityMap[severity];
+                const { background, text } = getSeverityColor(label);
+                return (
+                  <div style={{ display: 'flex', flexDirection: 'column', padding: '4px' }}>
+                    <EuiBadge key={severity} style={{ backgroundColor: background, color: text }}>
+                      {label}
+                    </EuiBadge>
+                  </div>
+                );
+              })}
+            </>
+          );
+
+          return (
+            <span
+              style={{
+                display: 'flex',
+                flexDirection: 'row',
+                alignItems: 'center',
+                gap: '4px',
+                whiteSpace: 'normal',
+                flexWrap: 'wrap',
+                width: '100%',
+              }}
+            >
+              {displayedSeverities}
+              {remainingCount > 0 && (
+                <EuiToolTip content={tooltipContent} position="top">
+                  <EuiBadge>{`+${remainingCount} more`}</EuiBadge>
+                </EuiToolTip>
+              )}
+            </span>
+          );
+        },
+      },
+      {
+        field: 'findingsSeverity',
+        name: 'Findings Severity',
+        sortable: true,
+        render: (findingsSeverity: string[]) => {
+          if (!findingsSeverity || findingsSeverity.length === 0) return DEFAULT_EMPTY_DATA;
+          const MAX_DISPLAY = 2;
+          const remainingCount =
+            findingsSeverity.length > MAX_DISPLAY ? findingsSeverity.length - MAX_DISPLAY : 0;
+          const displayedSeverities = findingsSeverity.slice(0, MAX_DISPLAY).map((severity) => {
+            const label = getSeverityLabel(severity);
+            const { background, text } = getSeverityColor(label);
+            return (
+              <EuiBadge key={severity} style={{ backgroundColor: background, color: text }}>
+                {label}
+              </EuiBadge>
+            );
+          });
+
+          const tooltipContent = (
+            <div
+              style={{
+                maxWidth: '300px',
+                maxHeight: '400px',
+                overflow: 'auto',
+              }}
+            >
+              {findingsSeverity.slice(MAX_DISPLAY).map((severity) => {
+                const label = getSeverityLabel(severity);
+                const { background, text } = getSeverityColor(label);
+                return (
+                  <div
+                    style={{
+                      display: 'flex',
+                      flexDirection: 'column',
+                      padding: '4px',
+                      width: '100%',
+                    }}
+                  >
+                    <EuiBadge key={severity} style={{ backgroundColor: background, color: text }}>
+                      {label}
+                    </EuiBadge>
+                  </div>
+                );
+              })}
+            </div>
+          );
+
+          return (
+            <span
+              style={{
+                display: 'flex',
+                flexDirection: 'row',
+                alignItems: 'center',
+                gap: '4px',
+                whiteSpace: 'normal',
+                flexWrap: 'wrap',
+                width: '100%',
+              }}
+            >
+              {displayedSeverities}
+              {remainingCount > 0 && (
+                <EuiToolTip content={tooltipContent} position="top">
+                  <EuiBadge>{`+${remainingCount} more`}</EuiBadge>
+                </EuiToolTip>
+              )}
+            </span>
+          );
+        },
+      },
+      {
+        field: 'actions',
+        name: 'Actions',
+        render: (_, correlationTableRow: CorrelationsTableData) => {
+          return (
+            <EuiToolTip content={'View details'}>
+              <EuiSmallButtonIcon
+                aria-label={'View details'}
+                data-test-subj={`view-details-icon`}
+                iconType={'inspect'}
+                onClick={() => {
+                  this.openTableFlyout(correlationTableRow);
+                }}
+              />
+            </EuiToolTip>
+          );
+        },
+      },
+    ];
+
+    const getRowProps = (item: any) => {
+      return {
+        'data-test-subj': `row-${item.id}`,
+        key: item.id,
+        className: 'euiTableRow',
+      };
+    };
+
+    const filteredTableData = this.getFilteredTableData(this.state.correlationsTableData);
+
+    return (
+      <>
+        {this.renderCorrelatedFindingsChart()}
+
+        <EuiInMemoryTable
+          items={filteredTableData}
+          rowProps={getRowProps}
+          columns={columns}
+          pagination={{
+            initialPageSize: 5,
+            pageSizeOptions: [5, 10, 20],
+          }}
+          responsive={true}
+          tableLayout="auto"
+        />
+      </>
+    );
+  };
+
+  private prepareGraphData = (correlationPairs: CorrelationFinding[][] | [any, any][]) => {
+    const createdEdges = new Set<string>();
+    const createdNodes = new Set<string>();
+    const graphData: CorrelationGraphData = {
+      graph: {
+        nodes: [],
+        edges: [],
+      },
+      events: {
+        click: this.onNodeClick,
+      },
+    };
+
+    correlationPairs.forEach((correlation: CorrelationFinding[]) => {
+      const possibleCombination1 = `${correlation[0].id}:${correlation[1].id}`;
+      const possibleCombination2 = `${correlation[1].id}:${correlation[0].id}`;
+
+      if (createdEdges.has(possibleCombination1) || createdEdges.has(possibleCombination2)) {
+        return;
+      }
+
+      if (!createdNodes.has(correlation[0].id)) {
+        this.addNode(graphData.graph.nodes, correlation[0]);
+        createdNodes.add(correlation[0].id);
+      }
+      if (!createdNodes.has(correlation[1].id)) {
+        this.addNode(graphData.graph.nodes, correlation[1]);
+        createdNodes.add(correlation[1].id);
+      }
+      this.addEdge(graphData.graph.edges, correlation[0], correlation[1]);
+      createdEdges.add(possibleCombination1);
+    });
+
+    return graphData;
+  };
+
+  private getCorrelationPairs = (correlatedFindings: any[]) => {
+    const pairs: [any, any][] = [];
+    for (let i = 0; i < correlatedFindings.length; i++) {
+      for (let j = i + 1; j < correlatedFindings.length; j++) {
+        pairs.push([correlatedFindings[i], correlatedFindings[j]]);
+      }
+    }
+    return pairs;
+  };
+
+  private renderTableFlyout = () => {
+    const { isFlyoutOpen, selectedTableRow, graphData } = this.state;
+
+    if (!isFlyoutOpen || !selectedTableRow) {
+      return null;
+    }
+
+    const findingsTableColumns: EuiBasicTableColumn<FlyoutTableData>[] = [
+      {
+        field: 'timestamp',
+        name: 'Time',
+        render: (timestamp: string) => new Date(timestamp).toLocaleString(),
+        sortable: true,
+      },
+      {
+        field: 'mitreTactic',
+        name: 'Mitre Tactic',
+        sortable: true,
+      },
+      {
+        field: 'detectionRule',
+        name: 'Detection Rule',
+        sortable: true,
+      },
+      {
+        field: 'severity',
+        name: 'Severity',
+        render: (ruleSeverity: string) => {
+          const severity = capitalizeFirstLetter(ruleSeverity) || DEFAULT_EMPTY_DATA;
+          const { background, text } = getSeverityColor(severity);
+
+          return (
+            <EuiBadge color={background} style={{ color: text }}>
+              {severity}
+            </EuiBadge>
+          );
+        },
+      },
+    ];
+
+    const flyoutTableData: FlyoutTableData[] = [];
+
+    selectedTableRow.correlatedFindings.map((correlatedFinding) => {
+      flyoutTableData.push({
+        timestamp: correlatedFinding.timestamp,
+        mitreTactic:
+          correlatedFinding.detectionRule.tags?.map((mitreTactic) => mitreTactic.value) || [],
+        detectionRule: correlatedFinding.detectionRule.name,
+        severity: correlatedFinding.detectionRule.severity,
+      });
+    });
+
+    return (
+      <EuiFlyout ownFocus onClose={this.closeTableFlyout} size="l" aria-labelledby="flyoutTitle">
+        <EuiFlyoutHeader hasBorder>
+          <EuiTitle size="m">
+            <h2 id="flyoutTitle">Correlation Details</h2>
+          </EuiTitle>
+          <EuiSpacer />
+          <EuiFlexGroup>
+            <EuiFlexItem>
+              <EuiText>
+                <p>
+                  <EuiTextColor color="subdued">Time</EuiTextColor>
+                  <br />
+                  {new Date(selectedTableRow.startTime).toLocaleString()}
+                </p>
+              </EuiText>
+            </EuiFlexItem>
+            <EuiFlexItem>
+              <EuiText>
+                <p>
+                  <EuiTextColor color="subdued">Correlation Rule</EuiTextColor>
+                  <br />
+                  {selectedTableRow.correlationRule}
+                </p>
+              </EuiText>
+            </EuiFlexItem>
+          </EuiFlexGroup>
+        </EuiFlyoutHeader>
+        <EuiFlyoutBody>
+          <div style={{ height: '100%', overflowY: 'auto' }}>
+            <EuiPanel>
+              <EuiTitle>
+                <h3>Correlated Findings</h3>
+              </EuiTitle>
+              {selectedTableRow.correlatedFindings && (
+                <CorrelationGraph
+                  loadingData={this.state.loadingGraphData}
+                  graph={graphData.graph}
+                  options={{
+                    ...graphRenderOptions,
+                    height: '300px',
+                    width: '100%',
+                    autoResize: true,
+                  }}
+                  events={graphData.events}
+                  getNetwork={this.setNetwork}
+                />
+              )}
+            </EuiPanel>
+            <EuiSpacer />
+            <EuiTitle size="xs">
+              <h3>Findings</h3>
+            </EuiTitle>
+            <EuiInMemoryTable
+              items={flyoutTableData || []}
+              columns={findingsTableColumns}
+              pagination={{
+                initialPageSize: 5,
+                pageSizeOptions: [5, 10, 20],
+              }}
+              sorting={true}
+            />
+            <EuiSpacer size="l" />
+            <EuiFlexGroup wrap responsive={false} gutterSize="m">
+              {Array.from(
+                new Set(
+                  selectedTableRow.correlatedFindings.flatMap(
+                    (finding) => finding.detectionRule.tags?.map((tag) => tag.value) || []
+                  )
+                )
+              ).map((tactic, i) => {
+                const link = `https://attack.mitre.org/techniques/${tactic
+                  .split('.')
+                  .slice(1)
+                  .join('/')
+                  .toUpperCase()}`;
+
+                return (
+                  <EuiFlexItem grow={false} key={i}>
+                    <EuiBadge color={'#DDD'}>
+                      <EuiLink href={link} target="_blank">
+                        {tactic}
+                      </EuiLink>
+                    </EuiBadge>
+                  </EuiFlexItem>
+                );
+              })}
+            </EuiFlexGroup>
+          </div>
+        </EuiFlyoutBody>
+      </EuiFlyout>
+    );
+  };
+
+  toggleView = () => {
+    this.setState((prevState) => ({
+      isGraphView: !prevState.isGraphView,
+    }));
+  };
 
   render() {
     const findingCardsData = this.state.specificFindingInfo;
@@ -529,6 +1420,7 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
                 Higher correlation score indicated stronger correlation.
               </EuiText>
               <EuiSpacer />
+
               {findingCardsData.correlatedFindings.map((finding, index) => {
                 return (
                   <>
@@ -574,6 +1466,7 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
           <EuiFlexItem>
             <EuiPanel>
               <EuiFlexGroup gutterSize="xs" wrap={false} justifyContent="flexEnd">
+                <EuiFlexItem grow={true}>{this.renderSearchBar()}</EuiFlexItem>
                 <EuiFlexItem grow={false}>
                   <EuiFilterGroup>
                     <FilterGroup
@@ -595,33 +1488,35 @@ export class Correlations extends React.Component<CorrelationsProps, Correlation
                     Reset filters
                   </EuiSmallButtonEmpty>
                 </EuiFlexItem>
-              </EuiFlexGroup>
-              <EuiSpacer />
-              <EuiFlexGroup
-                wrap={true}
-                gutterSize="m"
-                justifyContent="flexStart"
-                alignItems="center"
-              >
                 <EuiFlexItem grow={false}>
-                  <EuiText size="s">
-                    <strong>Severity:</strong>
-                  </EuiText>
+                  <EuiButtonGroup
+                    legend="View type"
+                    options={[
+                      {
+                        id: 'graph',
+                        label: 'Graph',
+                        iconType: 'visNetwork',
+                      },
+                      {
+                        id: 'table',
+                        label: 'Table',
+                        iconType: 'tableOfContents',
+                      },
+                    ]}
+                    idSelected={this.state.isGraphView ? 'graph' : 'table'}
+                    onChange={(id) => this.setState({ isGraphView: id === 'graph' })}
+                    buttonSize="s"
+                  />
                 </EuiFlexItem>
-                {ruleSeverity.map((sev, idx) => (
-                  <EuiFlexItem grow={false} key={idx}>
-                    <EuiText size="s">
-                      <EuiIcon type="dot" color={sev.color.background} /> {sev.value}
-                    </EuiText>
-                  </EuiFlexItem>
-                ))}
               </EuiFlexGroup>
-
               <EuiSpacer />
-              {this.renderCorrelationsGraph(this.state.loadingGraphData)}
+              {this.state.isGraphView
+                ? this.renderCorrelationsGraph(this.state.loadingGraphData)
+                : this.renderCorrelationsTable()}
             </EuiPanel>
           </EuiFlexItem>
         </EuiFlexGroup>
+        {this.renderTableFlyout()}
       </>
     );
   }
